[PDP-130] Adding new KafkaConsumer class
