+++ quintoandar_kafka/consumer.py
@@ -1,4 +1,6 @@
+import json
 import logging
+from retrying import retry
 
 from kafka.consumer.group import KafkaConsumer
 from quintoandar_kafka.idempotence_client import RedisIdempotenceClient
@@ -33,3 +35,52 @@ class KafkaIdempotentConsumer(KafkaSimpleConsumer):
             msg = super().__next__()
         self.idempotence_client.mark_consumed_message(msg.topic, msg)
         return msg
+
+
+class KafkaConsumer:
+    def __init__(
+        self, redis_host, redis_port, group_id, bootstrap_servers, topic, idempotent_key
+    ):
+        self.log = logging.getLogger(__name__)
+        self.consumer = self._connect(
+            redis_host, redis_port, group_id, bootstrap_servers, idempotent_key
+        )
+        self.consumer.subscribe(topic)
+
+    @retry(stop_max_attempt_number=10, wait_fixed=3000)
+    def _connect(
+        self, redis_host, redis_port, group_id, bootstrap_servers, idempotent_key
+    ):
+        return KafkaIdempotentConsumer(
+            redis_host=redis_host,
+            redis_port=redis_port,
+            idempotent_key=idempotent_key,
+            value_deserializer=self._deserializer,
+            group_id=group_id,
+            bootstrap_servers=bootstrap_servers,
+            auto_offset_reset="latest",
+        )
+
+    def _deserializer(self, m):
+        try:
+            return json.loads(m.decode("utf8"))
+        except TypeError as ex:
+            self.log.error("Failed to decode message: %s", ex, exc_info=True)
+            return {}
+
+    def read(self, event_processor_tuples):
+        for message in self.consumer:
+            if not message.value:
+                continue
+
+            self.log.info(message)
+
+            for (event, processor) in event_processor_tuples:
+                # skip
+                if message.value.get("eventName") != event:
+                    continue
+
+                try:
+                    processor(message.value)
+                except Exception as ex:
+                    self.log.error("Error processing message: %s", ex, exc_info=True)
